Sprint 4

+++ src/workspace/removeItemForm.js
@@ -3,18 +3,15 @@
  * from the inventory.
  */
 
-function removeItemFormModule(workbook=null, namespace="") {
-    if (workbook === null) {
-        workbook = SpreadsheetApp.getActiveSpreadsheet();
-    }
-
+function removeItemFormModule(workspace=null) {
+    workspace = Workspace.currentOr(workspace);
+    
     return new Component(
-        workbook,
-        namespace,
+        workspace,
         _removeItemFormNameFor,
-        (ns) => {
-            const form = _createRemoveItemForm(workbook, ns);
-            createSettings(workbook, ns).setRemoveItemForm(form);
+        () => {
+            const form = _createRemoveItemForm(workspace);
+            createSettings(workspace).setRemoveItemForm(form);
             return form;
         },
         _onRemoveItemFormSubmit
@@ -29,21 +26,18 @@ function _removeItemFormNameFor(namespace="") {
     return nameFor("Remove item", namespace);
 }
 
-function _createRemoveItemForm(workbook, namespace) {
-    const form = FormApp.create(_removeItemFormNameFor(namespace));
+function _createRemoveItemForm(workspace) {
+    const form = FormApp.create(_removeItemFormNameFor(workspace.namespace));
     form.setDescription("Remove an existing item from the inventory.");
+    form.setCollectEmail(true);
 
-    _populateRemoveItemForm(form, workbook, namespace);
+    _populateRemoveItemForm(form, workspace);
     
     return form;
 }
 
-function _populateRemoveItemForm(form, workbook=null, namespace="") {
-    if (workbook === null) {
-        workbook = SpreadsheetApp.getActiveSpreadsheet();
-    }
-
-    const service = createItemService(workbook, namespace);
+function _populateRemoveItemForm(form, workspace) {
+    const service = createItemService(workspace);
     const itemNames = service.getAllEntities().map(item => item.name);
     if (itemNames.length === 0) {
         itemNames.push("---");
@@ -60,26 +54,31 @@ function _onRemoveItemFormSubmit(event) {
      * appended after the timestamp. The last value in the array is the one we
      * want, so grab that one.
      */
-    const name = event.values.at(-1);
+    const name = lastNonEmpty(event.values);
+    console.log({
+        event: `Received remove item form from ${getEmailAddressFrom(event)}`,
+        itemName: name 
+    });
 
-    createItemService().remove(name);
-    Workspace.current().itemsChanged();
+    if (name) {
+        createItemService().remove(name);
+        Workspace.current().itemsChanged();
+    } else {
+        console.log("Name is falsey, so nothing to remove.");
+    }
 }
 
-function regenerateRemoveItemFormFor(workbook=null, namespace="") {
-    if (workbook === null) {
-        workbook = SpreadsheetApp.getActiveSpreadsheet();
-    }
-
-    const name = _removeItemFormNameFor(namespace);
-    const sheet = workbook.getSheetByName(name);
+function regenerateRemoveItemFormFor(workspace=null) {
+    workspace = Workspace.currentOr(workspace);
+    const name = _removeItemFormNameFor(workspace.namespace);
+    const sheet = workspace.workbook.getSheetByName(name);
     if (sheet === null) { // not generated yet
-        removeItemFormModule(workbook, namespace).setup();
+        removeItemFormModule(workspace).setup();
     } else {
         // remove old choices, repopulate
         const formUrl = sheet.getFormUrl();
         const form = FormApp.openByUrl(formUrl);
         form.getItems().forEach(item => form.deleteItem(item));
-        _populateRemoveItemForm(form, workbook, namespace);
+        _populateRemoveItemForm(form, workspace);
     }
 }
